requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
passlib
bcrypt==4.1.2
python-jose

